// src/app/models/counselor.ts
export interface Counselor {
    appointmentlink: string;
    credentials: string[];
    directoryid: string;
    email: string;
    firstname: string;
    insurance: string[];
    lastname: string;
    longdescription: string;
    memberships: string[];
    phone: string;
    practitionerid: string;
    shortdescription: string;
    titles: string;
}